# frontend/openapi/smart-meeting.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Meeting",
    "version": "0.1.0"
  },
  "paths": {
    "/api/update": {
      "get": {
        "summary": "Get Update",
        "description": "The latest version published, and how this installation gets it.",
        "operationId": "get_update",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Apply Update",
        "description": "Install the latest version, then restart: the server by itself in the browser version,\nthe desktop app restarts itself and its server.",
        "operationId": "apply_update",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/preferences": {
      "get": {
        "summary": "Get Preferences",
        "operationId": "get_preferences",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Preferences",
        "description": "Saved and applied at once: name and vocabulary for the next sentences, defaults for the next\nmeeting, the AI model (downloaded if missing).",
        "operationId": "update_preferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meetings": {
      "get": {
        "summary": "List Meetings",
        "description": "History, newest first; `q` searches titles, summaries and transcripts.",
        "operationId": "list_meetings",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeetingListItem"
                  },
                  "title": "Response List Meetings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Start Meeting",
        "operationId": "start_meeting",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartMeetingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/import": {
      "post": {
        "summary": "Import Meeting",
        "description": "Upload an audio or video file to transcribe and analyze. It is deleted once decoded.",
        "operationId": "import_meeting",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_meeting"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}": {
      "get": {
        "summary": "Get Meeting",
        "operationId": "get_meeting",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Meeting",
        "operationId": "update_meeting",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeetingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Meeting",
        "operationId": "delete_meeting",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/tags": {
      "put": {
        "summary": "Set Meeting Tags",
        "description": "Replaces the tags of a meeting (used to group the history).",
        "operationId": "set_meeting_tags",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/speakers": {
      "put": {
        "summary": "Rename Speaker",
        "description": "Names a speaker (\"Intervenant 2\" -> \"Paul\") in the whole meeting, during or after it.\nGiving the name of another speaker merges both (a voice wrongly split in two).",
        "operationId": "rename_speaker",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameSpeakerRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meetings/{meeting_id}/stop": {
      "post": {
        "summary": "Stop Meeting",
        "operationId": "stop_meeting",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/analyze": {
      "post": {
        "summary": "Analyze Meeting",
        "operationId": "analyze_meeting",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/ask": {
      "post": {
        "summary": "Ask Meeting",
        "description": "A question about the meeting (\"what do I have to do?\"), answered from its transcript by the\nlocal AI, during or after the meeting. Kept with the meeting.",
        "operationId": "ask_meeting",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskAnswer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meetings/{meeting_id}/report.md": {
      "get": {
        "summary": "Meeting Report",
        "operationId": "meeting_report",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/audio": {
      "delete": {
        "summary": "Delete Audio",
        "operationId": "delete_audio",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_import_meeting": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_meeting"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Preferences": {
        "properties": {
          "user_name": {
            "type": "string",
            "maxLength": 80,
            "title": "User Name",
            "default": "Moi"
          },
          "glossary": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Glossary",
            "default": []
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "auto"
          },
          "mic_device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mic Device"
          },
          "output_device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Device"
          },
          "keep_audio": {
            "type": "boolean",
            "title": "Keep Audio",
            "default": false
          },
          "room": {
            "type": "boolean",
            "title": "Room",
            "default": false
          },
          "ai_mode": {
            "type": "string",
            "enum": [
              "fast",
              "precise"
            ],
            "title": "Ai Mode",
            "default": "fast"
          },
          "ui_language": {
            "type": "string",
            "enum": [
              "fr",
              "en"
            ],
            "title": "Ui Language",
            "default": "fr"
          }
        },
        "type": "object",
        "title": "Preferences",
        "description": "Settings of the interface, applied without restart (see preferences.py)."
      },
...
      "StartMeetingRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "mic_device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mic Device"
          },
          "remote_device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Device"
          },
          "keep_audio": {
            "type": "boolean",
            "title": "Keep Audio",
            "default": false
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "auto"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Tags",
            "default": []
          },
          "room": {
            "type": "boolean",
            "title": "Room",
            "default": false
          }
        },
        "type": "object",
        "title": "StartMeetingRequest"
      },
...
      "UpdateInfo": {
        "properties": {
          "current": {
            "type": "string",
            "title": "Current"
          },
          "latest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "method": {
            "type": "string",
            "enum": [
              "git",
              "deb",
              "download"
            ],
            "title": "Method"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "updating": {
            "type": "boolean",
            "title": "Updating",
            "default": false
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "current",
          "latest",
          "available",
          "method",
          "url",
          "notes"
        ],
        "title": "UpdateInfo"
      },
...
      "UpdateResult": {
        "properties": {
          "restart": {
            "type": "string",
            "enum": [
              "server",
              "app"
            ],
            "title": "Restart"
          }
        },
        "type": "object",
        "required": [
          "restart"
        ],
        "title": "UpdateResult"
      },
...
    }
  }
}